Fix the evaluation of expressions when no keyword is given (= / do)
While those two prefixes really indicate that the following should be interpreted as an expression, the 'evaluate text field' option is also available when none of those two is given.
In this case, the two first characters of the actual command got cut, resulting in weird behaviour. This should be fixed by now.